locked editing and deleting articles for publishers

+++ app/Http/Controllers/Admin/ArticleController.php
@@ -41,12 +41,16 @@ class ArticleController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('ArticleController');
+
         $article = Article::query()->findOrFail($id);
         return view('admin.articles.edit', ['article' => $article]);
     }
 
     public function update(Request $request, $id)
     {
+        $this->authorize('ArticleController');
+
         $this->validate($request, [
             'title' => 'required|string|max:255',
             'file' => 'image|mimes:jpg,jpeg,png',
@@ -81,6 +85,8 @@ class ArticleController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('ArticleController');
+
         $article = Article::query()->findOrFail($id);
 
         $article->delete();
